+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -4,12 +4,11 @@ import com.sprint.mission.discodeit.dto.data.MessageDto;
 import com.sprint.mission.discodeit.dto.request.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.dto.request.MessageCreateRequest;
 import com.sprint.mission.discodeit.dto.request.MessageUpdateRequest;
-import com.sprint.mission.discodeit.dto.respond.PageResponse;
+import com.sprint.mission.discodeit.dto.response.PageResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.entity.base.BaseEntity;
 import com.sprint.mission.discodeit.mapper.MessageMapper;
 import com.sprint.mission.discodeit.mapper.PageResponseMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
@@ -21,10 +20,11 @@ import com.sprint.mission.discodeit.storage.BinaryContentStorage;
 import java.time.Instant;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,14 +32,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class BasicMessageService implements MessageService {
 
-  private final BinaryContentRepository binaryContentRepository;
-  private final BinaryContentStorage binaryContentStorage;
-  private final MessageMapper messageMapper;
   private final MessageRepository messageRepository;
-  private final PageResponseMapper pageResponseMapper;
   //
   private final ChannelRepository channelRepository;
   private final UserRepository userRepository;
+  private final MessageMapper messageMapper;
+  private final BinaryContentStorage binaryContentStorage;
+  private final BinaryContentRepository binaryContentRepository;
+  private final PageResponseMapper pageResponseMapper;
 
   @Transactional
   @Override
@@ -48,19 +48,13 @@ public class BasicMessageService implements MessageService {
     UUID channelId = messageCreateRequest.channelId();
     UUID authorId = messageCreateRequest.authorId();
 
-    if (!channelRepository.existsById(channelId)) {
-      throw new NoSuchElementException("Channel with id " + channelId + " does not exist");
-    }
-    if (!userRepository.existsById(authorId)) {
-      throw new NoSuchElementException("Author with id " + authorId + " does not exist");
-    }
-
     Channel channel = channelRepository.findById(channelId)
         .orElseThrow(
-            () -> new NoSuchElementException("Channel with id " + channelId + " not found"));
-
-    User user = userRepository.findById(authorId)
-        .orElseThrow(() -> new NoSuchElementException("Author with id " + authorId + " not found"));
+            () -> new NoSuchElementException("Channel with id " + channelId + " does not exist"));
+    User author = userRepository.findById(authorId)
+        .orElseThrow(
+            () -> new NoSuchElementException("Author with id " + authorId + " does not exist")
+        );
 
     List<BinaryContent> attachments = binaryContentCreateRequests.stream()
         .map(attachmentRequest -> {
@@ -70,9 +64,9 @@ public class BasicMessageService implements MessageService {
 
           BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
               contentType);
-          BinaryContent save = binaryContentRepository.save(binaryContent);
+          binaryContentRepository.save(binaryContent);
           binaryContentStorage.put(binaryContent.getId(), bytes);
-          return save;
+          return binaryContent;
         })
         .toList();
 
@@ -80,9 +74,10 @@ public class BasicMessageService implements MessageService {
     Message message = new Message(
         content,
         channel,
-        user,
+        author,
         attachments
     );
+
     messageRepository.save(message);
     return messageMapper.toDto(message);
   }
@@ -93,32 +88,25 @@ public class BasicMessageService implements MessageService {
     return messageRepository.findById(messageId)
         .map(messageMapper::toDto)
         .orElseThrow(
-            () -> new NoSuchElementException("Message with id " + messageId + " not found"))
-        ;
+            () -> new NoSuchElementException("Message with id " + messageId + " not found"));
   }
-
-//  @Transactional(readOnly = true)
-//  @Override
-//  public PageResponse<MessageDto> findAllByChannelId(UUID channelId, Pageable pageable) {
-//    return pageResponseMapper.fromPage(messageRepository
-//        .findAllByChannelId(channelId, pageable)
-//        .map(messageMapper::toDto));
-//  }
 
   @Transactional(readOnly = true)
   @Override
-  public PageResponse<MessageDto> findAllByChannelId(UUID channelId, Instant cursor,
+  public PageResponse<MessageDto> findAllByChannelId(UUID channelId, Instant createAt,
       Pageable pageable) {
-    if (cursor != null) {
-      return pageResponseMapper.fromPage(messageRepository
-          .findAllByChannelId(channelId, cursor, pageable)
-          .map(messageMapper::toDto));
-    } else {
-      return pageResponseMapper.fromPage(messageRepository
-          .findAllByChannelId(channelId, pageable)
-          .map(messageMapper::toDto));
+    Slice<MessageDto> slice = messageRepository.findAllByChannelIdWithAuthor(channelId,
+            Optional.ofNullable(createAt).orElse(Instant.now()),
+            pageable)
+        .map(messageMapper::toDto);
+
+    Instant nextCursor = null;
+    if (!slice.getContent().isEmpty()) {
+      nextCursor = slice.getContent().get(slice.getContent().size() - 1)
+          .createdAt();
     }
 
+    return pageResponseMapper.fromSlice(slice, nextCursor);
   }
 
   @Transactional
@@ -129,21 +117,15 @@ public class BasicMessageService implements MessageService {
         .orElseThrow(
             () -> new NoSuchElementException("Message with id " + messageId + " not found"));
     message.update(newContent);
-    messageRepository.save(message);
     return messageMapper.toDto(message);
   }
 
   @Transactional
   @Override
   public void delete(UUID messageId) {
-    Message message = messageRepository.findById(messageId)
-        .orElseThrow(
-            () -> new NoSuchElementException("Message with id " + messageId + " not found"));
-
-    message.getAttachments()
-        .stream()
-        .map(BinaryContent::getId)
-        .forEach(binaryContentRepository::deleteById);
+    if (!messageRepository.existsById(messageId)) {
+      throw new NoSuchElementException("Message with id " + messageId + " not found");
+    }
 
     messageRepository.deleteById(messageId);
   }
