+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -13,8 +13,6 @@ import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
-import java.io.IOException;
-import java.io.InputStream;
 import java.time.Instant;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -29,10 +27,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class BasicUserService implements UserService {
 
   private final UserRepository userRepository;
-  //
-  private final BinaryContentRepository binaryContentRepository;
   private final UserStatusRepository userStatusRepository;
   private final UserMapper userMapper;
+  private final BinaryContentRepository binaryContentRepository;
   private final BinaryContentStorage binaryContentStorage;
 
   @Transactional
@@ -56,61 +53,34 @@ public class BasicUserService implements UserService {
           byte[] bytes = profileRequest.bytes();
           BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
               contentType);
-          BinaryContent save = binaryContentRepository.save(binaryContent);
-          binaryContentStorage.put(save.getId(), bytes);
-          return save;
+          binaryContentRepository.save(binaryContent);
+          binaryContentStorage.put(binaryContent.getId(), bytes);
+          return binaryContent;
         })
         .orElse(null);
     String password = userCreateRequest.password();
+
     User user = new User(username, email, password, nullableProfile);
     Instant now = Instant.now();
     UserStatus userStatus = new UserStatus(user, now);
-    user.setStatus(userStatus);
 
-    User createdUser = userRepository.save(user);
-//    userStatusRepository.save(userStatus);
-
-    return userMapper.toDto(createdUser);
+    userRepository.save(user);
+    return userMapper.toDto(user);
   }
 
-  @Transactional(readOnly = true)
   @Override
   public UserDto find(UUID userId) {
-
     return userRepository.findById(userId)
         .map(userMapper::toDto)
-        .map(userDto -> {
-          InputStream inputStream = binaryContentStorage.get(userDto.profile().getId());
-          try {
-            byte[] bytes = inputStream.readAllBytes();
-            userDto.profile().setBytes(bytes);
-            return userDto;
-          } catch (IOException e) {
-            throw new RuntimeException(e);
-          }
-        })
         .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
   }
 
-  @Transactional(readOnly = true)
   @Override
   public List<UserDto> findAll() {
-    return userRepository.findAll()
+    return userRepository.findAllWithProfileAndStatus()
         .stream()
         .map(userMapper::toDto)
-        .map(userDto -> {
-          if (userDto.profile() == null) {
-            return userDto;
-          }
-          InputStream inputStream = binaryContentStorage.get(userDto.profile().getId());
-          try {
-            byte[] bytes = inputStream.readAllBytes();
-            userDto.profile().setBytes(bytes);
-            return userDto;
-          } catch (IOException e) {
-            throw new RuntimeException(e);
-          }
-        }).toList();
+        .toList();
   }
 
   @Transactional
@@ -131,37 +101,30 @@ public class BasicUserService implements UserService {
 
     BinaryContent nullableProfile = optionalProfileCreateRequest
         .map(profileRequest -> {
-          Optional.ofNullable(user.getProfile())
-              .map(BinaryContent::getId)
-              .ifPresent(binaryContentRepository::deleteById);
 
           String fileName = profileRequest.fileName();
           String contentType = profileRequest.contentType();
           byte[] bytes = profileRequest.bytes();
           BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
               contentType);
-          BinaryContent save = binaryContentRepository.save(binaryContent);
-          binaryContentStorage.put(save.getId(), bytes);
-          return save;
-
+          binaryContentRepository.save(binaryContent);
+          binaryContentStorage.put(binaryContent.getId(), bytes);
+          return binaryContent;
         })
         .orElse(null);
 
     String newPassword = userUpdateRequest.newPassword();
     user.update(newUsername, newEmail, newPassword, nullableProfile);
 
-    return userMapper.toDto(userRepository.save(user));
+    return userMapper.toDto(user);
   }
 
   @Transactional
   @Override
   public void delete(UUID userId) {
-    User user = userRepository.findById(userId)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
-
-    Optional.ofNullable(user.getProfile().getId())
-        .ifPresent(binaryContentRepository::deleteById);
-    userStatusRepository.deleteByUserId(userId);
+    if (userRepository.existsById(userId)) {
+      throw new NoSuchElementException("User with id " + userId + " not found");
+    }
 
     userRepository.deleteById(userId);
   }
